requirements.txt
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
blinker==1.5
boto3==1.24.38
botocore==1.27.38
click==8.1.3
decorator==5.1.1
executing==0.9.1
flake8==4.0.1
Flask==2.1.3
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.7.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
